@@ -1,10 +1,10 @@
 name: windows_documents
 description: Provides a document directory path for Windows
-version: 0.0.1
+version: 1.0.0
 author: Leon Tappe <ltappe@mail.upb.de>
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.20.0"
 
 dependencies:
